+++ Calculator/Gui/main.cpp
@@ -6,7 +6,8 @@ int main(int argc, char *argv[])
     QApplication app(argc, argv);
 
     MainWindow window;
-    window.setMinimumSize(320, 420);
+    window.setMinimumSize(520, 420);
+    window.resize(560, 460);
     window.show();
 
     return app.exec();